initial commit to 0.0.8 release

+++ src/download.py
@@ -1,47 +1,48 @@
 import os
-from tqdm import tqdm
 import asyncio
 import aiohttp
 import aiofiles
+from tqdm import tqdm
 
 from .core import BaseSession, Album
 
 
 class DownloadSession(BaseSession):
+
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.user = kwargs.get('user')
-        self.system = kwargs.get('system', False)
-        self.albums = self.get_all_albums()
+        self.system = kwargs.get('system') or 0
+        self.loop.run_until_complete(self.get_all_albums())
 
-    def get_all_albums(self):
-        response = self.api.photos.getAlbums(
+    async def get_all_albums(self):
+        response = await self.api.photos.getAlbums(
             owner_id=self.user,
             need_system=self.system
         )
 
-        return [Album(self.api, **item) for item in response['items']]
+        self.albums = [Album(self.api, **item) for item in response['items']]
 
     def get_album_by_id(self, id):
         return next((album for album in self.albums if album.id == id), None)
 
     def download_album(self, album):
+        self.loop.run_until_complete(album.get_photos())
         os.makedirs(os.path.join(self.path, album.title), exist_ok=True)
         path_to_album = os.path.join(self.path, album.title)
         path_url_pairs = tuple(
             (
-                os.path.join(path_to_album, photo.url.split('/')[-1]),
+                os.path.join(
+                    path_to_album,
+                    photo.url.split('/')[-1].replace('-', '')
+                ),
                 photo.url
-            ) for photo in album.get_photos()
+            ) for photo in album.photos
         )
 
-        self._download_async(album.title, path_url_pairs)
-
-    @staticmethod
-    def _download_async(title, path_url_pairs):
-        loop = asyncio.new_event_loop()
-        loop.run_until_complete(_download_many(title, path_url_pairs))
-        loop.close()
+        self.loop.run_until_complete(
+            _download_many(album.title, path_url_pairs)
+        )
 
 
 async def _download_many(title, path_url_pairs):
@@ -62,6 +63,9 @@ async def _download_many(title, path_url_pairs):
 
 
 async def _download_one(client, path, url):
+    if os.path.exists(path):
+        return
+
     async with client.get(url) as response:
         if response.status == 200:
             file = await aiofiles.open(path, 'wb')
